unittest: change macro expansion handling
Formerly, `check-macro-expansion` expands the form and compare it with
expected expansion. This is inconsistent with how other behaviours are
inspected and verified.
To make things consistent, now test "thunk" do all the inspection and
`check-*` do verification of actual behaviour later on.

`run-tests-1` isn't adopted for this change and unittest framework
doesn't work for now (these bits will be changed soon).