unittest: add support for `=> non-nil`
When `=> non-nil` appear in a `deftest` form, corresponding example
form is expected to return non-nil value.
Option `:compare COMPARE-FN`, if specified, is ignored for values
expected to be non-nil.
It can be one of multiple values.

    (deftest macroexpand-example-test ()
      "macroexpand returns expansion of form and true if form is
    actually expanded."
      (macroexpand '(setf x 1))
      => (setq x 1)
      => non-nil)

I initially planned to implement `=> non-nil` as entire return values,
instead of just one of values. IOW, the form is expected to return
non-nil value as its' primary value, and other values are not concerned.
Expected results `(:return :true)` and corresponding failure `(:results
:false)` is implemented for this purpose.
But treating it just specifying one of returned values make much more
sense, so they are removed now.