unittest: display test resluts buffer with `pop-to-buffer`
Now test results buffer is displayed via `pop-to-buffer` instead of
`set-buffer` (this differs when test results buffer is already displayed
in other window) at beginning of run.

Also, test results buffer is displayed at end of run as well if it is
not displayed, and scrolls to show summary line.